+++ Admin/src/main/java/com/example/admin/Domain/ImportGoodsDetail.java
@@ -14,5 +14,5 @@ public class ImportGoodsDetail {
 
     private int quantity;
 
-    private double price_Item;
+    private Long price_Item;
 }
