+++ Admin/src/main/java/com/example/admin/Domain/Order.java
@@ -20,11 +20,11 @@ public class Order {
 
     private LocalDateTime orderDateTime;
 
-    private double totalPrice;
+    private Long totalPrice;
 
-    private double discount;
+    private Long discount;
 
-    private double amount;
+    private Long amount;
 
     private Boolean confirmed;
 
@@ -38,7 +38,7 @@ public class Order {
 
     private boolean shipping;
 
-    private double shipPrice;
+    private Long shipPrice;
 
     private boolean paymentOnline;
