+++ Admin/src/main/java/com/example/admin/Repository/ImportGoodsRepository.java
@@ -16,5 +16,5 @@ public interface ImportGoodsRepository extends JpaRepository<ImportGoodsEntity, 
 
     List<ImportGoodsEntity> findImportGoodsEntitiesByTimeImportBetween(LocalDateTime startTime, LocalDateTime endTime);
     @Query("SELECT SUM(i.totalPrice) FROM ImportGoodsEntity i WHERE i.timeImport BETWEEN :startTime AND :endTime")
-    double getTotalImportAmountByTime(@Param("startTime") LocalDateTime startTime, @Param("endTime") LocalDateTime endTime);
+    Long getTotalImportAmountByTime(@Param("startTime") LocalDateTime startTime, @Param("endTime") LocalDateTime endTime);
 }
