+++ Admin/src/main/java/com/example/admin/Service/ImportGoodsService.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface ImportGoodsService {
     List<ImportGood> getImportGoodByTime(LocalDateTime startTime, LocalDateTime endTime);
     List<ImportGood> getAllImportGoods();
-    double getTotalImportAmountByTime(LocalDateTime startTime, LocalDateTime endTime);
+    Long getTotalImportAmountByTime(LocalDateTime startTime, LocalDateTime endTime);
 }
